🚨 Move templated code to header file
to make pybind11 happy
